박재혁/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.37.0
requests>=2.32.0
python-dotenv>=1.0.1
pydantic>=2.8.2
openai>=1.40.0
google-genai>=0.6.0
pillow>=10.4.0

